08-05-14

+++ HTML/comestibles.php
@@ -7,6 +7,7 @@
 	<title>Videoclub ORI</title>
 	<link rel="stylesheet" href="css/general.css">
 	<link rel="stylesheet" href="css/comestibles.css">
+	<link rel="icon" href="favicon.png" sizes="32x32" type="image/png">
 </head>
 <body>
 	
@@ -26,7 +27,7 @@
 				<article >
 					<h2>Comestibles</h2>
 					<?php
-						if(isset($_GET['eliminar'])){
+						if(isset($_GET['eliminar']) && isset($_SESSION['dni']) && $_SESSION['dni'] == "00000000A"){
 							$con = CrearConexionBD();
 							$id = $_GET['eliminar'];
 							$sql = "delete from comestibles where id_comestible=".$id;
@@ -37,7 +38,8 @@
 								}
 								echo '<div class="correcto"><p>Se ha eliminado el articulo</p></div>';
 							}else{
-								echo '<div class="incorrecto"><p>No se ha eliminado el articulo</p></div>';
+								echo '<div class="incorrecto"><p>No se ha eliminado el articulo</p></div>
+										<div class="incorrecto"><p>'.$con->errorInfo()[2].'</p></div>';
 							}
 							CerrarConexionBD($con);
 						}
@@ -54,10 +56,10 @@
 					<?php
 						$con = CrearConexionBD();
 						$res = $con->query("select id_comestible, nombre, to_char(precio, '990.99') from comestibles");
-						$bandera = 1==2;
+						$bandera = 0;
 						if(isset($_SESSION)){
-							if($_SESSION['dni']="00000000A"){
-								$bandera=1==1;
+							if($_SESSION['dni'] == "00000000A"){
+								$bandera = 1;
 							}
 						}
 						$cont = 0;
@@ -68,12 +70,16 @@
 										<input type="button" value="Eliminar" onClick=" window.location.href='."'comestibles.php?eliminar=".$fila[0]."'".'">
 										<input type="button" value="Modificar" onClick=" window.location.href='."'mod_comestible.php?id_comestible=".$fila[0]."'".'">
 										</td>';
-										$cont++;
+							}else{
+								echo '<td class="admin">
+									</td>';
 							}
 							
 							echo '		<td class="imagen"><img src="img_comestibles/'.$fila[0].'" /></td>
 									<td class="nombre">'.$fila[1].'</td>
 									<td class="precio">'.$fila[2].'€</td></tr>';
+
+							$cont++;
 								
 						}
 						CerrarConexionBD($con);
